unit쪽에서 기존에 직접 공격을 넘겨주던 방식에서 int? -> int  캐스팅후 넘겨주도록 변경

+++ Assets/Script/Ingame/UnitBuffHandler.cs
@@ -25,7 +25,7 @@ public class UnitBuffHandler : MonoBehaviour {
     private void Refresh() {
         placeMonster = GetComponent<PlaceMonster>();
         total_buffed_atk = placeMonster.unit.originalAttack;
-        total_buffed_hp = placeMonster.unit.hp;
+        total_buffed_hp = (int)placeMonster.unit.hp;
 
         origin_atk = total_buffed_atk;
         origin_hp = total_buffed_hp;
